niffler-tests-python/pyproject.toml
[project]
name = "niffler-py-st3"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pytest (==9.0.1)",
    "selenium (==4.38.0)",
    "allure-pytest (==2.15.2)",
    "python-dotenv (==1.2.1)",
    "mimesis (==18.0.0)",
    "requests (==2.32.5)"

]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
